# functions/swagger.yaml
openapi: 3.0.0
info:
  title: RMS API
  version: 1.0.0
  description: Restaurant Management System API
servers:
  - url: https://api-d4o6tbc5fq-uc.a.run.app/
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /doors:
    get:
      summary: Retrieve a list of all doors
      security:
        - bearerAuth: []
      tags:
        - Doors
      responses:
        '200':
          description: A list of doors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Door'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a new door
      security:
        - bearerAuth: []
      tags:
        - Doors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Door'
      responses:
        '201':
          description: Door created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The ID of the newly created door
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /doors/{doorId}:
    get:
      summary: Retrieve a door by ID
      security:
        - bearerAuth: []
      tags:
        - Doors
      parameters:
        - in: path
          name: doorId
          required: true
          schema:
            type: string
          description: The ID of the door to retrieve
      responses:
        '200':
          description: The door details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Door'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Door not found
    put:
      summary: Update a door by ID
      security:
        - bearerAuth: []
      tags:
        - Doors
      parameters:
        - in: path
          name: doorId
          required: true
          schema:
            type: string
          description: The ID of the door to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Door'
      responses:
        '200':
          description: Door updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Door'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Door not found
    delete:
      summary: Delete a door by ID
      security:
        - bearerAuth: []
      tags:
        - Doors
      parameters:
        - in: path
          name: doorId
          required: true
          schema:
            type: string
          description: The ID of the door to delete
      responses:
        '200':
          description: Door deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Door not found
...
  /opening-hours/{id}:
    get:
      summary: Retrieve opening hours by ID
      tags:
        - Opening Hours
      description: Retrieve the opening hours for a specific ID.
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the opening hours to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: A successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  openingHours:
                    type: string
        '404':
          description: Opening hours not found
    put:
      summary: Update opening hours
      tags:
        - Opening Hours
      description: Update existing opening hours.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the opening hours to update.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                day:
                  type: string
                startTime:
                  type: string
                endTime:
                  type: string
      responses:
        '200':
          description: Opening hours updated successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Opening hours not found
    delete:
      summary: Delete opening hours
      tags:
        - Opening Hours
      description: Delete existing opening hours.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the opening hours to delete.
          schema:
            type: string
      responses:
        '200':
          description: Opening hours deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Opening hours not found
  /opening-hours:
    get:
      summary: Retrieve all opening hours
      tags:
        - Opening Hours
      description: Retrieve all opening hours.
      responses:
        '200':
          description: A successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    openingHours:
                      type: string
    post:
      summary: Create new opening hours
      tags:
        - Opening Hours
      description: Create new opening hours.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                day:
                  type: string
                startTime:
                  type: string
                endTime:
                  type: string
      responses:
        '201':
          description: Opening hours created successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...